requirements.txt
django_debug_toolbar==1.6
django_multiselectfield==0.1.5
Django==1.11.2
wagtail==1.10.1
django-extensions>=1.7.5
django-contact-form==1.3
psycopg2>=2.6.2
django-dbbackup==3.1.3
django-dbbackup-ui==0.1.5
django-autocomplete-light==3.2.1
django-charsleft-widget==0.1.6
django-embed-video==1.1.2
django-compressor==2.1.1
csscompressor==0.9.4
django-redis==4.8.0
django-environ==0.4.3